feat: :rocket: init
