+++ src/modules/d.js
@@ -1,0 +1,6 @@
+'use strict'
+
+import * as c2e from './cjsToEsm.mjs'
+// import { unknownObject } from './cjsToEsm.mjs'
+
+console.log(c2e.unknownObject)
